Implemented improved native aek targets for testing
