+++ stackgvm/source/junk/aek_2.cpp
@@ -9,76 +9,7 @@
 //   * Moved the specular calculation to the point where the ray definitely
 //     intersects a sphere.
 
-#include <cstdlib>
-#include <cstdio>
-#include <cmath>
-
-typedef int int32;
-typedef float float32;
-
-///////////////////////////////////////////////////////////////////////////////
-
-// Vector class. Replaced operator overloads with
-struct vec3 {
-    float32 x, y, z;
-
-    // default constructor
-    vec3() { }
-
-    // constructor
-    vec3(float32 a, float32 b, float32 c) {
-        x = a;
-        y = b;
-        z = c;
-    }
-};
-
-const vec3 camera_dir(
-    -6.0f, -16.0f, 0.0f
-);
-
-const vec3 focal_point(
-    17.0f, 16.0f, 8.0f
-);
-
-const vec3 normal_up(
-    0.0f, 0.0f, 1.0f
-);
-
-const vec3 sky_rgb(
-    0.7f, 0.6f, 1.0f
-);
-
-const vec3 floor_red_rgb(
-    3.0f, 1.0f, 1.0f
-);
-
-const vec3 floor_white_rgb(
-    3.0f, 3.0f, 3.0f
-);
-///////////////////////////////////////////////////////////////////////////////
-
-#ifdef NO_PASS_BY_REF
-typedef const vec3 cvr3;
-#else
-typedef const vec3& cvr3;
-#endif
-
-///////////////////////////////////////////////////////////////////////////////
-
-// Data - bitvector of sphere locations, 1 means a sphere, 0 means empty space
-//
-int32 data[] = {
-    247570, // 0111100011100010010
-    280596, // 1000100100000010100
-    280600, // 1000100100000011000
-    249748, // 0111100111110010100
-    18578,  // 0000100100010010010
-    18577,  // 0000100100010010001
-    231184, // 0111000011100010000
-    16,     // 0000000000000010000
-    16      // 0000000000000010000
-};
+#include "aek.hpp"
 
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -98,63 +29,6 @@ void init_spheres() {
         }
     }
     std::fprintf(stderr, "num_spheres:%d\n", num_spheres);
-}
-
-// Sum two vec3
-static inline vec3 vec3_add(cvr3 v1, cvr3 v2) {
-    return vec3(
-        v1.x + v2.x,
-        v1.y + v2.y,
-        v1.z + v2.z
-    );
-}
-
-// Subtract two vec3
-static inline vec3 vec3_sub(cvr3 v1, cvr3 v2) {
-    return vec3(
-        v1.x - v2.x,
-        v1.y - v2.y,
-        v1.z - v2.z
-    );
-}
-
-// Scale a vec3 by a float
-static inline vec3 vec3_scale(cvr3 v, float32 s) {
-    return vec3(
-        v.x * s,
-        v.y * s,
-        v.z * s
-    );
-}
-
-// Get a normalised vec3
-static inline vec3 vec3_normalize(cvr3 v) {
-    return vec3_scale(v, (1.0 / std::sqrt(
-        (v.x * v.x) +
-        (v.y * v.y) +
-        (v.z * v.z)
-    )));
-}
-
-// Get the dot product of two vec3
-static inline float32 dot(cvr3 v1, cvr3 v2) {
-    return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
-}
-
-// Get the cross product for two vec3
-static inline vec3 vec3_cross(cvr3 v1, cvr3 v2) {
-    return vec3(
-        v1.y * v2.z - v1.z * v2.y,
-        v1.z * v2.x - v1.x * v2.z,
-        v1.x * v2.y - v1.y * v2.x
-    );
-}
-
-static const float32 invRM = 1.0 / RAND_MAX;
-
-// Get a random number in the range 0.0 - 1.0
-static inline float32 frand() {
-    return invRM * rand();
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -177,14 +51,14 @@ int32 trace(cvr3 origin, cvr3 direction, float32& distance, vec3& normal) {
     // Check if trace maybe hits a sphere
     for (int i = 0; i<num_spheres; i++) {
         vec3 p = vec3_sub(
-        origin,
-        spheres[i] // Sphere coordinate
+            origin,
+            spheres[i] // Sphere coordinate
         );
 
         float32
-        b = dot(p, direction),
-        eye_offset = dot(p, p) - 1.0f,
-        q = b * b - eye_offset
+            b = dot(p, direction),
+            eye_offset = dot(p, p) - 1.0f,
+            q = b * b - eye_offset
         ;
         if (q > 0.0f) {
             float32 sphere_distance = -b - std::sqrt(q);
@@ -216,8 +90,8 @@ vec3 sample(cvr3 origin, cvr3 direction) {
         gradient *= gradient;
         gradient *= gradient;
         return vec3_scale(
-        sky_rgb, // Blueish sky colour
-        gradient
+            sky_rgb, // Blueish sky colour
+            gradient
         );
     }
 
@@ -244,14 +118,14 @@ vec3 sample(cvr3 origin, cvr3 direction) {
     }
 
     // Hit the floor plane
-    if (material & 1) {
+    if (material == 1) {
         intersection = vec3_scale(intersection, 0.2f);
         return vec3_scale(
             (
                 // Compute check colour based on the position
                 (int32) (ceil(intersection.x) + ceil(intersection.y)) & 1 ?
-                floor_red_rgb : // red
-                floor_white_rgb   // white
+                    floor_red_rgb : // red
+                    floor_white_rgb   // white
             ),
             (lambertian * 0.2f + 0.1f)
         );
@@ -283,7 +157,7 @@ vec3 sample(cvr3 origin, cvr3 direction) {
 ///////////////////////////////////////////////////////////////////////////////
 
 // Main
-int32 main() {
+int main() {
     int image_size = 512;
     printf("P6 %d %d 255 ", image_size, image_size);
