+++ stackgvm/source/junk/aek_single.cpp
